+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Added `ExFirebaseAuth.Cookie` to support for verifying session cookies.
+
 ## 0.5.1
 
 - Set default expiry on mocked token to 1 hour from utc now.